TokenPrice Observation phase
* Add GetFeeQuoterTokenUpdates in onchain_prices_reader.go
* Observe feedPrices, feeQuoterUpdates, fDestChain and current timestamp (UTC)

Note: We're observing Fchain, although we already observe it in merkleroot processor. This is because it's more straight forward if each processor takes care of what they need. For example the token price processor don't need all the Fs for all chains, just the destination fChain. Also reduces complexity of the PluginProcessor design as sending an extra parameter will require changing the current interface. If we see that it's affecting performance in the future then we can change the interface.

Signed-off-by: asoliman <abdelrahman.soliman@smartcontract.com>