commit plugin - support batch sizes greater than one (#131)
So far commit plugin did not support batch sizes greater than 1, meaning that the reported sequence number ranges were always of size 1 e.g. chain1: [123, 123].

In this PR I use the ccipReader to observe the onRamp sequence numbers to increase the batch sizes.
Since every onRamp exists on a different chain we cannot use batch calls, goroutines are spawned instead for each source chain using errgroup.
If for a specific source chain we cannot get the onRamp sequence number, onRamp sequence numbers observation stops.
Commit plugin should limit the amount of data it can process in a round. In a scenario when we have too many pending messages, e.g. [123, 10000] we limit this ranges up to a specific configured batch size which is set to a default of 256.

Limits can be configured on plugin initialization.