rmn - cursing checks and reverts prevention (#321)
Commit plugin cursing checks and reverts prevention.

While observing offRamp sequence numbers, the oracle will make a call to RMNRemote to get the cursed source chains and global/dest cursing information.

    In case of a global or dest curse, no seqNums are observered, leading to no merkleRoot processor observation.
    In case of a source chain curse, no seqNums are observed for this particular chain leading to merkle roots not observed for this source chain.

While in the tranmission protocol, the oracle makes a similar check to validate whether the report will cause a revert or not. If the report will revert onChain due to cursed chain(s) it's not transmitted.

Similar to the offRamp logic, if the report only contains chain fee and token prices while on a global/dest curse it is transmitted.