pkg/logutil: ctx-aware logging, add seq nr to all commit logs (#458)
* pkg/logutil: add ocr seq nr to all commit logs

Ticket: https://smartcontract-it.atlassian.net/browse/CCIP-4877

In order to build more readable logs, we include the OCR sequence
number in the context that is provided by libocr. This means that
all downstream components, i.e the ccip reader, processors, etc.
will be able to extract the sequence number and include it as a
field in their loggers.

* some more renames

* move constants + types to top

* rm large field from log