Fix exec plugin issue related to filtering executed msgs from reports (#509)
Replace ExecutedMessageRanges function with ExecutedMessages which returns a slice of the executed messages in the provided sequence numbers range. Update all the execution plugin related logic that called this function. This leads to making the calculations very simple without having to deal with complex sequence number ranges of 1 element each.