commit: Only attempt to fetch latestMsgSeqNum from enabled source chains
We found this when testing with Solana: the default setup requires two
EVM chains: a home chain and a feeds chain. In the solana setup the feeds
chain has a configured relayer but no lane connecting it (isEnabled: false)

This does add an extra read to the Observation call, perhaps the config
should be cached?