commit/merkleroot: fix report transmission checks (#773)
There seems to have been two issues with the report transmission checks:
* We didn't set RootsToReport correctly in the Outcome where we were
checking for the report transmission. This caused the next round to believe
that the report was successfully transmitted.
* We don't check for ReportInFlight in the Reports function. I believe
this would cause us to include the same merkle roots in the report again
even though they may have been transmitted successfully.

This commit fixes both of these issues.