Exec Plugin - Update consensus algorithm on nonces (#776)
Improving the algorithm for selecting the nonces among multiple observations.

The existing algorithm was selecting nonces according to "how much votes" a specific nonce received.
But this is problematic since we might have a constant stream of msgs coming from the same sender and we might have consensus issues. E.g. we don't reach the threshold because every oracle observes a different nonce.

The algorithm in this PR is similar to what we do in the commit plugin for sequence numbers where we sort the sequence numbers ascending and select the f-th one (f+1 since arrays are zero indexed).