Limit number of reports and messages in exec plugin GetMessages phase [CCIP-5647]  (#812)
As Exec plugin has 3 states, the second state GetMessages was getting all possible messages (with 1K reports) without doing much filtering other than if the observation size exceeded max size using type_optimizer - deprecated in this PR -.

The new algorithm does filter based on the encoding size as well but it's more lenient at the GetMessages phase. Usually exec plugin is able to execute 10-20 msgs at a time so it doesn't make sense to keep on observing - in theory - thousands of messages every round to only come in the third round and remove most of them to send only 10 msgs.

We're setting a couple of lenient maxMsgPerObs and maxObsSize that we're sure can keep up with without complicating the algorithm.

Also introduced logic to check for size limit which used iteratively for each message added to the observation (ditching the old way of reading all and then removing messages later)