commit/chainfee: better logs, always include FChain and TimestampNow
When reading this code I noticed that the uniqueChains terminology
didn't really make much sense; it seems that we only report fee
components and native token prices when we have both of those values
simultaneously for a particular chain. So I named these variables more
clearly and added them to the "observed fee components" log to make it
clear what chains were filtered out due to lack of token prices.

I also added FChain and TimestampNow to the empty observation returned
in the case where there are no native token prices for any chain to
avoid the "no consensus for fDestChain" log that would be misleading in
that scenario, since it simply didn't get observed appropriately.