(commit from remote subtree): Proxying tokens through TP ATA (#9)
Because EVM does not have access to the token receiver's ATA, the CCTP
message cannot be sent from EVM with the right mint_recipient. Thus, the
workaround is to always send the CCTP funds to an intermediary TP-owned
ATA, and then have the TP transfer the funds to the actual token
receiver.