ccipcapability: move interface impls
Move over the interface implementations, from the core
smartcontractkit/chainlink repo, to a new go module within this repo
called ccipcapability.

This means that we can easily add an implementation for an interface
that is newly defined without breaking builds elsewhere. Better
co-location of interface and implementation also means less back and
forth across repos which has historically been a big time sink.

As of this commit, the ccipcapability module is importing
smartcontractkit/chainlink but the goal is to not do that. However, the
main sticking point is evmrelaytypes, the package that defines the EVM
chainreader configuration struct. Ideally this moves to chainlink-evm at
some point soon (see e.g.
https://github.com/smartcontractkit/chainlink/pull/16578) and we'd only
have to import chainlink-evm instead.

Some remaining things in chainlink would be the launcher and
oraclecreator packages, which currently heavily depend on chainlink
packages. They first need to be abstracted away from those so that we
can safely extract them.