Add TxHash to Message
TxHash was recently added to the Sequence type in contract_reader. In this
commit we bump the version of chainlink-common to access this new field and
use it to populate the TxHash field of Message. The forthcoming CCTPv2
attestation logic needs TxHash to fetch attestations for a Message's USDC
transfers.